+++ backend/api/mapcreator/points/points.service.js
@@ -1,4 +1,5 @@
-const database = require("#sql/database.js");
+const database = require("#mapcreator/points/points.queries.js");
+const { getConnection, insertImage, updateImagePath, deleteImageById, getPointImage } = require("#sql/database.js");
 const AppError = require("#utils/app-error.js");
 const fs = require("fs/promises");
 const path = require("path");
@@ -23,7 +24,7 @@ async function updatePoint(userId, pointID, pointData, file) {
 
         await assertUserOwnsPoint(userId, pointID);
 
-        dbConnection = await database.getConnection();
+        dbConnection = await getConnection();
         await dbConnection.beginTransaction();
 
         let pointInfo = await database.getPointInfo(pointID);
@@ -51,7 +52,7 @@ async function updatePoint(userId, pointID, pointData, file) {
         }
 
         if (file) {
-            let oldImageInfo = await database.getPointImage(pointID);
+            let oldImageInfo = await getPointImage(pointID);
 
             let imageData;
             try {
@@ -59,7 +60,7 @@ async function updatePoint(userId, pointID, pointData, file) {
             } catch (err) {
                 throw new AppError(ERRORS.COMMON.IMAGE_PROCESSING_ERROR, 422);
             }
-            let newImageId = await database.insertImage(dbConnection, imageData.width, imageData.height, "pending");
+            let newImageId = await insertImage(dbConnection, imageData.width, imageData.height, "pending");
             let gameMapID = pointInfo.game_maps_id;
             let mapID = pointInfo.map_id;
 
@@ -83,7 +84,7 @@ async function updatePoint(userId, pointID, pointData, file) {
 
             let dbPath = path.join(relativeDestDir, processedImagePaths.targetFileName);
 
-            await database.updateImagePath(dbConnection, newImageId, dbPath);
+            await updateImagePath(dbConnection, newImageId, dbPath);
 
             let updateImageSuccess = await database.updatePointImage(dbConnection, pointID, newImageId);
             if (!updateImageSuccess) {
@@ -91,7 +92,7 @@ async function updatePoint(userId, pointID, pointData, file) {
             }
 
             if (oldImageInfo) {
-                let deleteSuccess = await database.deleteImageById(dbConnection, oldImageInfo.image_id);
+                let deleteSuccess = await deleteImageById(dbConnection, oldImageInfo.image_id);
                 if (!deleteSuccess) {
                     throw new AppError(ERRORS.POINT.OLD_IMAGE_DELETION_FAILED, 500);
                 }
@@ -143,6 +144,7 @@ async function createPoint(userId, mapID, pointData, file) {
         const gameMapID = await database.getGameMapIdByMapId(mapID);
         if (!gameMapID) {
             throw new AppError(ERRORS.COMMON.UNEXPECTED_ERROR, 500);
+            // TODO: throw new AppError(ERRORS.GAMEMAP.NOT_FOUND, 404);
         }
 
         let imageData;
@@ -152,7 +154,7 @@ async function createPoint(userId, mapID, pointData, file) {
             throw new AppError(ERRORS.COMMON.IMAGE_PROCESSING_ERROR, 422);
         }
 
-        dbConnection = await database.getConnection();
+        dbConnection = await getConnection();
         await dbConnection.beginTransaction();
 
         let existingPoints = await database.getPointOnMapByCoordinates(dbConnection, mapID, uCoordinate, vCoordinate);
@@ -160,7 +162,7 @@ async function createPoint(userId, mapID, pointData, file) {
             throw new AppError(ERRORS.POINT.ALREADY_EXISTS, 409);
         }
 
-        let imageId = await database.insertImage(dbConnection, imageData.width, imageData.height, "pending");
+        let imageId = await insertImage(dbConnection, imageData.width, imageData.height, "pending");
 
         let newPointId = await database.insertPoint(dbConnection, mapID, uCoordinate, vCoordinate, northDirection, imageId);
 
@@ -186,7 +188,7 @@ async function createPoint(userId, mapID, pointData, file) {
 
         let dbPath = path.join(relativeDestDir, processedImagePaths.targetFileName);
 
-        await database.updateImagePath(dbConnection, imageId, dbPath);
+        await updateImagePath(dbConnection, imageId, dbPath);
 
         await dbConnection.commit();
 
@@ -219,13 +221,13 @@ async function deletePoint(userId, pointID) {
             throw new AppError(ERRORS.POINT.NOT_FOUND, 404);
         }
 
-        let oldImageInfo = await database.getPointImage(pointID);
+        let oldImageInfo = await getPointImage(pointID);
 
-        dbConnection = await database.getConnection();
+        dbConnection = await getConnection();
         await dbConnection.beginTransaction();
 
         if (oldImageInfo && oldImageInfo.image_id) {
-            let successImageDeletion = await database.deleteImageById(dbConnection, oldImageInfo.image_id);
+            let successImageDeletion = await deleteImageById(dbConnection, oldImageInfo.image_id);
             if (!successImageDeletion) {
                 throw new AppError(ERRORS.POINT.IMAGE_DELETION_FAILED, 500);
             }
