+++ backend/sql/__mocks__/database.js
@@ -17,16 +17,7 @@ const mockDatabase = {
     insertImage: jest.fn(),
     updateImagePath: jest.fn().mockResolvedValue(true),
     deleteImageById: jest.fn().mockResolvedValue(true),
-    getPointsOnMap: jest.fn(),
-    getGameMapIdByMapId: jest.fn(),
-    getPointOnMapByCoordinates: jest.fn(),
-    insertPoint: jest.fn(),
-    getPointInfo: jest.fn(),
     getPointImage: jest.fn(),
-    deletePointById: jest.fn().mockResolvedValue(true),
-    updatePointCoordinates: jest.fn().mockResolvedValue(true),
-    updatePointNorthDirection: jest.fn().mockResolvedValue(true),
-    updatePointImage: jest.fn().mockResolvedValue(true),
     insertConnection: jest.fn(),
 
     getGameMapComments: jest.fn(),
