+++ backend/test/jest.setup.js
@@ -5,3 +5,4 @@ jest.mock("#root/auth.js");
 jest.mock("#gamemaps/shared/middlewares/gamemaps.middleware.js");
 jest.mock("#mapcreator/connections/connections.queries.js");
 jest.mock("#mapcreator/maps/maps.queries.js");
+jest.mock("#mapcreator/points/points.queries.js");
